+++ SourceCode/gamesystem/character/NormalSKill.cpp
@@ -1,6 +1,6 @@
 #include "NormalSkill.h"
 #include "imgui.h"
-void NormalSkill::Create(const std::string name, int ID, float latency, float invocatingTime, float restraintTime, int rarity, int popRate, float damege, XMFLOAT2 range, XMFLOAT2 distance)
+void NormalSkill::Create(const std::string name, int ID, int latency, int invocatingTime, int restraintTime, int rarity, int popRate, float damege, XMFLOAT2 range, XMFLOAT2 distance)
 {
 	this->name = name;
 	this->ID = ID;
